refactor: Java 17 업그레이드 및 도메인 모델 구조 리팩토링
