+++ core/domain/src/main/java/com/afternote/core/domain/model/DailyQuestionModel.kt
@@ -1,0 +1,13 @@
+package com.afternote.core.domain.model
+
+data class DailyQuestionAnswerItem(
+    val dailyQuestionAnswerId: Long,
+    val question: String,
+    val answer: String,
+    val recordDate: String,
+)
+
+data class ReceiverDailyQuestionsResult(
+    val items: List<DailyQuestionAnswerItem>,
+    val hasNext: Boolean,
+)
