+++ core/domain/src/main/java/com/afternote/core/domain/model/MindRecordModel.kt
@@ -1,0 +1,14 @@
+package com.afternote.core.domain.model
+
+data class ReceiverMindRecordItem(
+    val recordId: Long,
+    val type: String,
+    val titleOrQuestion: String,
+    val contentOrAnswer: String,
+    val recordDate: String,
+)
+
+data class ReceiverMindRecordsResult(
+    val items: List<ReceiverMindRecordItem>,
+    val hasNext: Boolean,
+)
