+++ core/domain/src/main/java/com/afternote/core/domain/model/UserProfileModel.kt
@@ -1,0 +1,8 @@
+package com.afternote.core.domain.model
+
+data class UserProfile(
+    val name: String,
+    val email: String,
+    val phone: String?,
+    val profileImageUrl: String?,
+)
