+++ core/domain/src/main/java/com/afternote/core/domain/model/setting/ReceiverModel.kt
@@ -1,0 +1,19 @@
+package com.afternote.core.domain.model.setting
+
+data class ReceiverListItem(
+    val receiverId: Long,
+    val name: String,
+    val relation: String,
+    val mindRecordDeliveryEnabled: Boolean = true,
+)
+
+data class ReceiverDetail(
+    val receiverId: Long,
+    val name: String,
+    val relation: String,
+    val phone: String?,
+    val email: String?,
+    val dailyQuestionCount: Int,
+    val timeLetterCount: Int,
+    val afterNoteCount: Int,
+)
