+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/CreateSocialInput.kt
@@ -1,4 +1,4 @@
-package com.kuit.afternote.feature.afternote.domain.model
+package com.afternote.feature.afternote.domain.model
 
 data class CreateSocialInput(
     val title: String,
