+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/GetAfternotesInput.kt
@@ -1,5 +1,6 @@
-package com.kuit.afternote.feature.afternote.domain.model
+package com.afternote.feature.afternote.domain.model
 
+// AfternoteRepository.getAfternotes() 인자
 data class GetAfternotesInput(
     val category: String?,
     val page: Int,
