--- feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/model/PagedAfternotes.kt
@@ -1,14 +1,0 @@
-package com.kuit.afternote.feature.afternote.domain.model
-
-import com.afternote.feature.afternote.domain.model.Item
-
-/**
- * One page of afternotes from GET /afternotes.
- *
- * @param items Items for this page
- * @param hasNext Whether more pages are available
- */
-data class PagedAfternotes(
-    val items: List<Item>,
-    val hasNext: Boolean,
-)
