+++ feature/afternote/domain/src/main/java/com/afternote/feature/afternote/domain/usecase/CreateGalleryAfternoteUseCase.kt
@@ -1,6 +1,6 @@
 package com.afternote.feature.afternote.domain.usecase
 
-import com.kuit.afternote.feature.afternote.domain.model.CreateGalleryInput
+import com.afternote.feature.afternote.domain.model.CreateGalleryInput
 import com.kuit.afternote.feature.afternote.domain.repository.AfternoteRepository
 import javax.inject.Inject
